# docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: Simple chat server (Gin + Gorm)
  title: Bubble Chat API
  version: 0.1.0
paths:
...
  /api/blacklist:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: List blacklist
      tags:
      - blacklist
    post:
      consumes:
      - application/json
      parameters:
      - description: User to block
        in: body
        name: body
        required: true
        schema:
          properties:
            userId:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Add to blacklist
      tags:
      - blacklist
  /api/blacklist/{userId}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Remove from blacklist
      tags:
      - blacklist
...
  /api/friends:
    delete:
      parameters:
      - description: Friend User ID
        in: query
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Remove friend
      tags:
      - friends
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: List friends
      tags:
      - friends
  /api/friends/{id}/nickname:
    put:
      consumes:
      - application/json
      - application/json
      description: 设置好友备注
      parameters:
      - description: '{nickname}'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      - description: '{nickname}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 不是好友关系
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 用户不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Set friend nickname
      tags:
      - friends
      - friends
  /api/friends/{userId}/nickname:
    put:
      consumes:
      - application/json
      - application/json
      description: 设置好友备注
      parameters:
      - description: Friend User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: '{nickname}'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      - description: Friend user ID
        in: path
        name: userId
        required: true
        type: integer
      - description: '{nickname}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 不是好友关系
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 用户不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Set friend nickname
      tags:
      - friends
      - friends
  /api/friends/accept:
    post:
      consumes:
      - application/json
      parameters:
      - description: '{fromName}'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Accept friend request
      tags:
      - friends
  /api/friends/requests:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: List incoming friend requests
      tags:
      - friends
    post:
      consumes:
      - application/json
      parameters:
      - description: '{toName}'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Send friend request
      tags:
      - friends
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"